use int64 instead of int for elements_appearing_once

+++ epi/element_appearing_once/solution.go
@@ -1,6 +1,6 @@
 package element_appearing_once
 
-func FindElementAppearsOnce(a []int) int {
+func FindElementAppearsOnce(a []int64) int64 {
 	// TODO - Add your code here
 	return 0
 }
